Link reseve button on details page to reserve page
