+++ src/styles/details.module.css
@@ -1,4 +1,4 @@
-.contianer {
+.container {
   height: 100vh;
   display: flex;
   align-items: center;
@@ -10,7 +10,7 @@
   justify-content: center;
   align-items: center;
   gap: 20rem;
-  margin-top: -15rem;
+  /* margin-top: -15rem; */
 }
 
 .detailsText {
@@ -20,13 +20,32 @@
   margin-right: -10rem;
 }
 
-.contianer button {
-  width: 150px;
-  height: 50px;
+.container button {
   background-color: #96bf03;
   border: none;
   color: #fff;
+  font-weight: 600;
+}
+
+.container a {
+  text-decoration: none;
+}
+
+.reserveBtn {
+  color: #fff;
+  background-color: #96bf03;
+  width: 150px;
+  height: 50px;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
   border-radius: 25px;
-  font-weight: 500;
+  padding: 0 1rem;
   margin-top: 33rem;
+}
+
+.logo {
+  font-size: 1.5rem;
+  font-weight: 100;
+  opacity: 0.8;
 }
